+++ include/output.h
@@ -8,7 +8,7 @@ bool output_secondary_avail();
 void output_write(char *msg);
 void output_writeInt(int i);
 void output_putc(char c);
-void output_printHex(int n);
+void output_printHex(uintptr_t n);
 void output_changeToFB();
 void kprintf(char *format,...);
 void putc(char c);
